Fix ECONNRESET crash and add tui/collab player scenarios
- Fix unhandled ECONNRESET: add error handlers to all pipe chains in
  Vite proxy (HTTP, WebSocket upgrade, video streaming) and add
  process-level safety net for transient socket errors
- Add tui-terminals.scenario.ts and collab.scenario.ts using
  defineScenario() so they appear in the player selector
- Add s.options() to ScenarioBuilder for declaring session-level
  options (e.g. layout: "row") that both runScenario and the player
  executor respect
- Eliminate code duplication: basic-ui.test.ts, tui-terminals.test.ts,
  and collab.test.ts now import from their .scenario.ts counterparts
  via runScenario(descriptor)
